+++ src/components/Header/index.js
@@ -1,0 +1,19 @@
+import React from "react";
+import "./about.css";
+
+function About() {
+  return (
+
+    <div className="masthead container d-flex h-100 align-items-center">
+      <div className="mx-auto text-center">
+        <h1 className="mx-auto my-0 text-uppercase headerAutoBongs ">AutoBongs</h1>
+        <h2 className="text-white-50 mx-auto mt-2 mb-5">The Ultimate Monthly Subscription for Glass Collectors </h2>
+        <a href="pages/login.html" class="btn btn-primary js-scroll-trigger ">Get Started</a>
+      </div>
+    </div>
+
+
+  );
+}
+
+export default About;
