--- src/components/What/index.js
@@ -1,32 +1,0 @@
-import React from "react";
-import "./what.css";
-
-function What() {
-  return (
-
-    <section id="about" className="about-section text-center">
-    <div className="container">
-      <div className="row">
-        <div className="col-lg-8 mx-auto ">
-          <!-- <h2 className="text-white mb-4">What is AutoBongs?</h2> -->
-
-
-          <h4 className="text-white-50">Autobongs is the premiere monthly subscription box for the fine water
-            pipe
-            enthusiast.<br><br>
-            The service is comprised of 3 tiers of increasing quality and price.
-            Each tier is curated according to your preferences, hand picked and arranged just for you. <br><br>
-
-            We strive to deliver the most quality bang for your buck. Each box contains a varety of water pipes
-            and
-            accesorries with quality and quantity increasing with each tier.</h4>
-        </div>
-      </div>
-    </div>
-  </section>
-
-
-  );
-}
-
-export default What;
